+++ src/main/resources/Tables.sql
@@ -5,7 +5,7 @@ DROP TABLE SELLERS IF EXISTS;
 CREATE TABLE SELLERS (name varchar(255) primary key);
 
 CREATE TABLE PRODUCT (
-productId long,
+productId numeric,
 productName varchar(255),
  productPrice double,
  sellerName varchar(255) references SELLERS(name));
